==> src/05_github-networks/02_international-networks/03_country-networks/02_ctry-nets-mod-viz.Rmd <==
---
title: "Country Network Analysis"
output: html_document
---

```{r loading edgelist data, warning=FALSE, echo=FALSE}
rm(list = ls())
# install.packages(c("tidyverse", "igraph", "visNetwork", "bc3net", 

# load packages 
for (pkg in c("tidyverse", "data.table", "R.utils", "RPostgreSQL", "igraph", 
              "cowplot", "maditr", "lubridate", "countrycode")) {library(pkg, character.only = TRUE)}

# connect to postgresql to get data (in rivanna)
conn <- dbConnect(drv = PostgreSQL(), 
                  dbname = "sdad", 
                  host = "10.250.124.195", 
                  port = 5432, 
                  user = Sys.getenv("db_userid"), 
                  password = Sys.getenv("db_pwd"))

# query the bipartite edgelist data from github data  
sna_intl_ctry_nodelist <- dbGetQuery(conn, "SELECT country, users, repos, commits, additions, deletions
                                            FROM gh.sna_intl_ctry_nodelist_full")

# query the bipartite edgelist data from github data  
sna_intl_ctry_edgelist <- dbGetQuery(conn, "SELECT country1, country2, ctrycode1, ctrycode2, repo_wts
                                            FROM gh.sna_intl_ctry_edgelist_0819")

# disconnect from postgresql
dbDisconnect(conn)
```

```{r, fig.width=11}
# filter out the omitted_geographies 
omitted_geographies <- c("Multiple Countries", "Europe", "Asia", "The Americas", "Africa")
`%notin%` <- Negate(`%in%`)
sna_intl_ctry_nodelist <- sna_intl_ctry_nodelist %>% 
  filter(country %notin% omitted_geographies) 

sna_intl_ctry_edgelist_full <- sna_intl_ctry_edgelist %>% 
  select(country1, country2, repo_wts) %>% 
  mutate(weight = repo_wts) %>% select(-repo_wts)
  
intl_regular_network <- graph.data.frame(sna_intl_ctry_edgelist_full, directed = FALSE)

# first, we create one network object where the multiples are combined and loops remain 
# "multiples combined" means that combinations like albania-usa and usa-albania are collapsed
# i validated that the edge weights are correct for the collapsed multiples 
intl_ctry_network <- simplify(graph.data.frame(sna_intl_ctry_edgelist_full, directed = FALSE), 
                              edge.attr.comb = igraph_opt("edge.attr.comb"),
                              remove.loops = FALSE)

# next, we create an object that collapses multiples but eliminates the loops 
intl_ctry_network_noloops <- simplify(graph.data.frame(sna_intl_ctry_edgelist_full, directed = FALSE), 
                              edge.attr.comb = igraph_opt("edge.attr.comb"),
                              remove.loops = TRUE) 

# let's take a look at the full network with loops 
plot(intl_ctry_network, 
     edge.color="#a9a9a9", 
     edge.curved=0.1,
     edge.label=FALSE,
     vertex.size=8,
     vertex.color="#E57200", 
     vertex.label.color="#232D4B",
     layout=layout_with_kk)

``` 

This plot shows what we would expect. It looks as if (almost?) every country has some type of collaboration. But, in classic igraph fashion, the graph is (almost?) unbearable to look at. Let's export the node/edgelist into Gephi first and then we conduct break the network down over time to see what the dynamics look like. Before we do that, however, we want to perform use the Newman modularity algorithm on this network since Gephi's default Louvain algorithm is stochastic and harder to replicate for some plotting we do later. 

```{r}

# add the community membership to the nodelist 
intl_ctry_nodelist_gephi <- data.frame(id = c(1:(igraph::vcount(intl_ctry_network))), country = igraph::V(intl_ctry_network)$name)
intl_ctry_nodelist_gephi <- intl_ctry_nodelist_gephi %>% 
  inner_join(sna_intl_ctry_nodelist, by = "country") %>% 
  rename(region = country)
fstgrdy <- fastgreedy.community(intl_ctry_network)
intl_ctry_nodelist_gephi$fstgrdy_comm <- fstgrdy$membership

#intl_ctry_nodelist_gephi %>% filter(fstgrdy_comm != 1 & fstgrdy_comm != 2)

intl_ctry_nodelist_gephi$country_code <- countrycode(intl_ctry_nodelist_gephi$region, 
            origin = 'country.name', 
            destination = 'iso2c') 

intl_ctry_nodelist_gephi <- intl_ctry_nodelist_gephi %>% 
  mutate(country_code = ifelse(test = str_detect(string = region, 
                               pattern = "Kosovo"), 
                               yes = "XK", no = country_code))

# recode user ranges into categories  
intl_ctry_nodelist_gephi$recoded_comm <- cut(intl_ctry_nodelist_gephi$fstgrdy_comm, 
                                             breaks=c(0,1,2,3,4,100), 
                                             labels=c("1", "2", "3", "4", "5"))

# graph the communities on the world map to see if they need consolidating
map_data("world") %>% 
  mutate(country_code = countrycode(region, 
                                    origin = 'country.name', 
                                    destination = 'iso2c')) %>% 
  left_join(intl_ctry_nodelist_gephi,  
            by = "country_code") %>% 
  ggplot() +
  theme_minimal() +
  geom_polygon(aes(x = long, y = lat, 
                   fill = recoded_comm, 
                   group = group), 
                   color = "black") + 
  coord_fixed(1.2) +
  guides(fill=FALSE) + 
  scale_fill_manual(values = c("#E57200", # dark red    
                               "#628ed8", # light blue
                               "#990000", # orange 
                               "#232D4B", # dark blue
                               "#eaaa31"  # yellow
                               )) +
  theme(plot.title = element_text(size = 13, hjust = 0.5), 
        plot.caption = element_text(size = 10, vjust = 8, hjust = 0.85),
        axis.line=element_blank(),axis.text.x=element_blank(),
        axis.text.y=element_blank(),axis.ticks=element_blank(),
        axis.title.x=element_blank(), axis.title.y=element_blank()) +
  labs(title= " Modularity Groupings for Country-Country \n Collaboration Networks (GitHub, 2008-2019)",
       caption = "Note: Groupings Calculated with Loops in Network")

```

```{r}

# add the community membership to the nodelist 
sna_intl_ctry_nodelist_noloops <- data.frame(id = c(1:(igraph::vcount(intl_ctry_network_noloops))), 
                                             country = igraph::V(intl_ctry_network_noloops)$name)
sna_intl_ctry_nodelist_noloops <- sna_intl_ctry_nodelist_noloops %>% 
  inner_join(sna_intl_ctry_nodelist, by = "country") %>% 
  rename(region = country)
fstgrdy <- fastgreedy.community(intl_ctry_network_noloops)
sna_intl_ctry_nodelist_noloops$fstgrdy_comm <- fstgrdy$membership

#sna_intl_ctry_nodelist_noloops %>% filter(fstgrdy_comm != 1 & fstgrdy_comm != 2)

sna_intl_ctry_nodelist_noloops$country_code <- countrycode(sna_intl_ctry_nodelist_noloops$region, 
            origin = 'country.name', 
            destination = 'iso2c') 

sna_intl_ctry_nodelist_noloops <- sna_intl_ctry_nodelist_noloops %>% 
  mutate(country_code = ifelse(test = str_detect(string = region, 
                               pattern = "Kosovo"), 
                               yes = "XK", no = country_code))

# recode user ranges into categories  
#intl_ctry_nodelist_gephi$recoded_noloops <- cut(sna_intl_ctry_nodelist_noloops$fstgrdy_comm, 
#                                             breaks=c(0,1,2,100), 
#                                             labels=c("1", "2", "3"))

# graph the communities on the world map to see if they need consolidating
map_data("world") %>% 
  mutate(country_code = countrycode(region, 
                                    origin = 'country.name', 
                                    destination = 'iso2c')) %>% 
  left_join(intl_ctry_nodelist_gephi,  
            by = "country_code") %>% 
  ggplot() +
  theme_minimal() +
  geom_polygon(aes(x = long, y = lat, 
                   fill = fstgrdy_comm, 
                   group = group), 
                   color = "black") + 
  coord_fixed(1.2) +
  guides(fill=FALSE) + 
  #scale_fill_manual(values = c("#628ed8", # light blue
  #                             #"#E57200", # orange
  #                             "#232D4B" # dark blue
  #                             )) +
  theme(plot.title = element_text(size = 13, hjust = 0.5), 
        plot.caption = element_text(size = 10, vjust = 8, hjust = 0.87),
        axis.line=element_blank(),axis.text.x=element_blank(),
        axis.text.y=element_blank(),axis.ticks=element_blank(),
        axis.title.x=element_blank(), axis.title.y=element_blank()) +
  labs(title= " Modularity Groupings for Country-Country \n Collaboration Networks (GitHub, 2008-2019)",
       caption = "Note: Groupings Calculated without Loops")

```

```{r}

# we need to recode the country column to id for gephi 
intl_ctry_nodelist_gephi <- intl_ctry_nodelist_gephi %>% 
  rename(id = region) %>% 
  select(id, country_code, everything())

intl_ctry_nodelist_gephi

# we need to recode the country columns to source, target for gephi
intl_ctry_edgelist_gephi <- sna_intl_ctry_edgelist_full %>% 
  rename(source = country1, target = country2)

intl_ctry_edgelist_gephi

# export the full network nodelist and edgelist 
setwd("~/oss-data/intl-ctry-nets-cum")
write_csv(intl_ctry_nodelist_gephi, "intl_ctry_nodelist_gephi.csv")
write_csv(intl_ctry_edgelist_gephi, "intl_ctry_edgelist_gephi.csv") 

```

# Creating Country-to-Country Network Visuals in Gephi 


```{r}

library("HiveR")

?HiveR
test_edgelist <- sna_intl_ctry_edgelist %>% 
  select(-ctrycode1, -ctrycode2) %>% 
  rename(lab1 = country1, lab2 = country2, weight = repo_wts)

class(test_edgelist)
class(edges)

test <- edge2HPD(test_edgelist, axis.cols = NULL)


edges <- data.frame(lab1 = LETTERS[c(1:8, 7)],
					lab2 = LETTERS[c(2:4, 1:3, 4, 2, 2)],
					weight = c(1, 1, 2, 2, 3, 1, 2, 3, 1))
					
td <- edge2HPD(edge_df = edges, desc = "Test of edge2HPD")
plotptd.out <- sumHPD(td, plot.list = TRUE)



gAdj <- get.adjacency(intl_ctry_network, type = "upper", edges = FALSE, names = TRUE, sparse = FALSE)
hive1 <- adj2HPD(gAdj, type = "2D")
hive2 <- mineHPD(hive1, option = "rad <- tot.edge.count")
hive3 <- mineHPD(hive2, option = "axis <- source.man.sink")
hive4 <- mineHPD(hive3, option = "remove zero edge")
plotHive(hive4, method = "abs", bkgnd = "white", axLabs = c("source", "hub", "sink"), axLab.pos = 1)





############################################################################################
# Create a graph. Use simplify to ensure that there are no duplicated edges or self loops
gD <- intl_ctry_network

# Print number of nodes and edges
# vcount(gD)
# ecount(gD)

# Calculate some node properties and node similarities that will be used to illustrate
# different plotting abilities

# Calculate degree for all nodes
degAll <- degree(gD, v = V(gD), mode = "all")

# Calculate betweenness for all nodes
betAll <- betweenness(gD, v = V(gD), directed = FALSE) / (((vcount(gD) - 1) * (vcount(gD)-2)) / 2)
betAll.norm <- (betAll - min(betAll))/(max(betAll) - min(betAll))

node.list <- data.frame(name = V(gD)$name, degree = degAll, betw = betAll.norm)

# Calculate Dice similarities between all pairs of nodes
dsAll <- similarity.dice(gD, vids = V(gD), mode = "all")

# Calculate edge weight based on the node similarity
F1 <- function(x) {data.frame(V4 = dsAll[which(V(gD)$name == as.character(x$V1)), which(V(gD)$name == as.character(x$V2))])}
dataSet.ext <- ddply(dataSet, .variables=c("V1", "V2", "V3"), function(x) data.frame(F1(x)))

rm(degAll, betAll, betAll.norm, F1)
############################################################################################
#Determine node/edge color based on the properties

# Calculate node size
# We'll interpolate node size based on the node betweenness centrality, using the "approx" function
# And we will assign a node size for each node based on its betweenness centrality
approxVals <- approx(c(0.5, 1.5), n = length(unique(node.list$bet)))
nodes_size <- sapply(node.list$bet, function(x) approxVals$y[which(sort(unique(node.list$bet)) == x)])
node.list <- cbind(node.list, size = nodes_size)
rm(approxVals, nodes_size)

# Define node color
# We'll interpolate node colors based on the node degree using the "colorRampPalette" function from the "grDevices" library
library("grDevices")
# This function returns a function corresponding to a collor palete of "bias" number of elements
F2 <- colorRampPalette(c("#F5DEB3", "#FF0000"), bias = length(unique(node.list$degree)), space = "rgb", interpolate = "linear")
# Now we'll create a color for each degree
colCodes <- F2(length(unique(node.list$degree)))
# And we will assign a color for each node based on its degree
nodes_col <- sapply(node.list$degree, function(x) colCodes[which(sort(unique(node.list$degree)) == x)])
node.list <- cbind(node.list, color = nodes_col)
rm(F2, colCodes, nodes_col)

# Assign visual attributes to edges using the same approach as we did for nodes
F2 <- colorRampPalette(c("#FFFF00", "#006400"), bias = length(unique(dataSet.ext$V4)), space = "rgb", interpolate = "linear")
colCodes <- F2(length(unique(dataSet.ext$V4)))
edges_col <- sapply(dataSet.ext$V4, function(x) colCodes[which(sort(unique(dataSet.ext$V4)) == x)])
dataSet.ext <- cbind(dataSet.ext, color = edges_col)
rm(F2, colCodes, edges_col)

############################################################################################
# Assign nodes to axes

# Randomly
nodeAxis <- sample(3, nrow(node.list), replace = TRUE )
node.list <- cbind(node.list, axis = nodeAxis)
rm(nodeAxis)

############################################################################################
#Create a hive plot

source("mod.edge2HPD.R")

hive1 <- mod.edge2HPD(edge_df = dataSet.ext[, 1:2], edge.weight = dataSet.ext[, 3], edge.color = dataSet.ext[, 5], node.color = node.list[,c("name", "color")], node.size = node.list[,c("name", "size")], node.radius = node.list[,c("name", "degree")], node.axis = node.list[,c("name", "axis")])
#sumHPD(hive1)

hive2 <- mineHPD(hive1, option = "remove zero edge")

plotHive(hive2, method = "abs", bkgnd = "white",  axLab.pos = 1)



````
