==> src/05_github-networks/02_international-networks/03_country-networks/14_ctry-nets-usa-fraction.Rmd <==
---
title: "Create Edgelists"
output: html_document
---

```{r}
# connect to postgresql to get data (in rivanna)
conn <- dbConnect(drv = PostgreSQL(), 
                  dbname = "sdad", 
                  host = "10.250.124.195", 
                  port = 5432, 
                  user = Sys.getenv("db_userid"), 
                  password = Sys.getenv("db_pwd"))

# query the bipartite edgelist data from github data  
ctry_edgelist <- dbGetQuery(conn, "SELECT country1, country2, repo_wts
                                   FROM gh.sna_intl_ctry_edgelist_0819")

analysis_year <- "0819"

# disconnect from postgresql
dbDisconnect(conn)


# convert edgelist to network 
ctry_edgelist <- ctry_edgelist %>% 
  select(country1, country2, repo_wts) %>% 
  rename(from = country1, to = country2, weight = repo_wts) %>% 
  group_by(from, to) %>% 
  summarize(weight = sum(weight)) %>% 
  arrange(-weight)
  
ctry_network <- simplify(graph.data.frame(ctry_edgelist, directed = FALSE), 
                         remove.loops = FALSE, 
                         edge.attr.comb = igraph_opt("edge.attr.comb"))
is_weighted(ctry_network)

ctry_edgelist <- melt(as.matrix(as_adjacency_matrix(ctry_network, type = "lower", attr = "weight", sparse = T)))

ctry_edgelist <- ctry_edgelist %>% 
  rename(from = Var1, to = Var2, weight = value) %>% 
  arrange(-weight)

```

```{r}

top10 <- "Australia|Canada|China|France|Germany|India|Japan|Netherlands|United Kingdom|United States"
accidents <- "Hong Kong SAR China|Macau SAR China|Caribbean Netherlands|British Indian Ocean Territory"

usa_total <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "United States" | to == "United States") %>% 
  summarize(weight = sum(weight))

usa_vector <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "United States" | to == "United States") %>% 
  filter(grepl(top10, from)) %>% 
  filter(!grepl(accidents, from)) %>%
  filter(grepl(top10, to)) %>% 
  filter(!grepl(accidents, to)) %>%
  mutate(fraction = round(weight / usa_total$weight*100, 2)) %>% 
  arrange(-fraction) %>% 
  select(from, fraction) %>% 
  rename(country = from, USA = fraction)
usa_vector

```

```{r}

uk_total <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "United Kingdom" | to == "United Kingdom") %>% 
  summarize(weight = sum(weight))

uk_vector <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "United Kingdom" | to == "United Kingdom") %>% 
  filter(grepl(top10, from)) %>% filter(!grepl(accidents, from)) %>%
  filter(grepl(top10, to)) %>% filter(!grepl(accidents, to)) %>%
  mutate(fraction = round(weight / uk_total$weight*100, 2)) %>% 
  arrange(-fraction) %>% 
  select(from, fraction) %>% 
  rename(country = from, UK = fraction)

uk_vector$country[1] <- "United States"

fraction_data <- uk_vector %>% full_join(usa_vector); fraction_data

```

```{r}

china_total <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "China" | to == "China") %>% 
  summarize(weight = sum(weight))

china_vector <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "China" | to == "China") %>% 
  filter(grepl(top10, from)) %>% filter(!grepl(accidents, from)) %>%
  filter(grepl(top10, to)) %>% filter(!grepl(accidents, to)) %>%
  mutate(fraction = round(weight / china_total$weight*100, 2)) %>% 
  arrange(-fraction)

temp <- china_vector %>% 
  slice(1:4) %>% 
  mutate(country = to) %>% 
  select(country, fraction)

fraction_data <- china_vector %>% 
  slice(5:10) %>% 
  mutate(country = from) %>% 
  select(country, fraction) %>% 
  bind_rows(temp) %>% 
  arrange(-fraction) %>% 
  rename(China = fraction) %>% 
  right_join(fraction_data, by = "country"); rm(temp)

fraction_data
```

```{r}

germany_total <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "Germany" | to == "Germany") %>% 
  summarize(weight = sum(weight))

germany_vector <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "Germany" | to == "Germany") %>% 
  filter(grepl(top10, from)) %>% filter(!grepl(accidents, from)) %>%
  filter(grepl(top10, to)) %>% filter(!grepl(accidents, to)) %>%
  mutate(fraction = round(weight / germany_total$weight*100, 2)) %>% 
  arrange(-fraction)

temp <- germany_vector %>% 
  slice(1:2) %>% 
  mutate(country = to) %>% 
  select(country, fraction)

fraction_data <- germany_vector %>% 
  slice(3:10) %>% 
  mutate(country = from) %>% 
  select(country, fraction) %>% 
  bind_rows(temp) %>% 
  arrange(-fraction) %>% 
  rename(Germany = fraction) %>% 
  right_join(fraction_data, by = "country"); rm(temp)

fraction_data
```

```{r}

australia_total <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "Australia" | to == "Australia") %>% 
  summarize(weight = sum(weight))

australia_vector <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "Australia" | to == "Australia") %>% 
  filter(grepl(top10, from)) %>% filter(!grepl(accidents, from)) %>%
  filter(grepl(top10, to)) %>% filter(!grepl(accidents, to)) %>%
  mutate(fraction = round(weight / australia_total$weight*100, 2)) %>% 
  arrange(-fraction)

temp <- australia_vector %>% 
  slice(1:7) %>% 
  mutate(country = to) %>% 
  select(country, fraction)

fraction_data <- australia_vector %>% 
  slice(8:10) %>% 
  mutate(country = from) %>% 
  select(country, fraction) %>% 
  bind_rows(temp) %>% 
  arrange(-fraction) %>% 
  rename(Australia = fraction) %>% 
  right_join(fraction_data, by = "country"); rm(temp)

fraction_data
```
```{r}

canada_total <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "Canada" | to == "Canada") %>% 
  summarize(weight = sum(weight))

canada_vector <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "Canada" | to == "Canada") %>% 
  filter(grepl(top10, from)) %>% filter(!grepl(accidents, from)) %>%
  filter(grepl(top10, to)) %>% filter(!grepl(accidents, to)) %>%
  mutate(fraction = round(weight / canada_total$weight*100, 2)) %>% 
  arrange(-fraction)

temp <- canada_vector %>% 
  slice(1:4) %>% 
  mutate(country = to) %>% 
  select(country, fraction)

fraction_data <- canada_vector %>% 
  slice(5:10) %>% 
  mutate(country = from) %>% 
  select(country, fraction) %>% 
  bind_rows(temp) %>% 
  arrange(-fraction) %>% 
  rename(Canada = fraction) %>% 
  right_join(fraction_data, by = "country"); rm(temp)

fraction_data
```
```{r}

france_total <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "France" | to == "France") %>% 
  summarize(weight = sum(weight))

france_vector <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "France" | to == "France") %>% 
  filter(grepl(top10, from)) %>% filter(!grepl(accidents, from)) %>%
  filter(grepl(top10, to)) %>% filter(!grepl(accidents, to)) %>%
  mutate(fraction = round(weight / france_total$weight*100, 2)) %>% 
  arrange(-fraction)

temp <- france_vector %>% 
  slice(1:6) %>% 
  mutate(country = to) %>% 
  select(country, fraction)

fraction_data <- france_vector %>% 
  slice(7:10) %>% 
  mutate(country = from) %>% 
  select(country, fraction) %>% 
  bind_rows(temp) %>% 
  arrange(-fraction) %>% 
  rename(France = fraction) %>% 
  right_join(fraction_data, by = "country"); rm(temp)

fraction_data
```
```{r}

india_total <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "India" | to == "India") %>% 
  summarize(weight = sum(weight))

india_vector <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "India" | to == "India") %>% 
  filter(grepl(top10, from)) %>% filter(!grepl(accidents, from)) %>%
  filter(grepl(top10, to)) %>% filter(!grepl(accidents, to)) %>%
  mutate(fraction = round(weight / india_total$weight*100, 2)) %>% 
  arrange(-fraction)

temp <- india_vector %>% 
  slice(1:7) %>% 
  mutate(country = to) %>% 
  select(country, fraction)

fraction_data <- india_vector %>% 
  slice(8:10) %>% 
  mutate(country = from) %>% 
  select(country, fraction) %>% 
  bind_rows(temp) %>% 
  arrange(-fraction) %>% 
  rename(India = fraction) %>% 
  right_join(fraction_data, by = "country"); rm(temp)

fraction_data
```
```{r}

japan_total <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "Japan" | to == "Japan") %>% 
  summarize(weight = sum(weight))

japan_vector <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "Japan" | to == "Japan") %>% 
  filter(grepl(top10, from)) %>% filter(!grepl(accidents, from)) %>%
  filter(grepl(top10, to)) %>% filter(!grepl(accidents, to)) %>%
  mutate(fraction = round(weight / japan_total$weight*100, 2)) %>% 
  arrange(-fraction)

temp <- japan_vector %>% 
  slice(1:9) %>% 
  mutate(country = to) %>% 
  select(country, fraction)

fraction_data <- japan_vector %>% 
  slice(10) %>% 
  mutate(country = from) %>% 
  select(country, fraction) %>% 
  bind_rows(temp) %>% 
  arrange(-fraction) %>% 
  rename(Japan = fraction) %>% 
  right_join(fraction_data, by = "country"); rm(temp)

fraction_data
```
```{r}

netherlands_total <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "Netherlands" | to == "Netherlands") %>% 
  summarize(weight = sum(weight))

netherlands_vector <- ctry_edgelist %>% 
  filter(weight != 0) %>% 
  filter(from == "Netherlands" | to == "Netherlands") %>% 
  filter(grepl(top10, from)) %>% filter(!grepl(accidents, from)) %>%
  filter(grepl(top10, to)) %>% filter(!grepl(accidents, to)) %>%
  mutate(fraction = round(weight / netherlands_total$weight*100, 2)) %>% 
  arrange(-fraction)

fraction_data <- netherlands_vector %>% 
  mutate(country = to) %>% 
  select(country, fraction) %>% 
  arrange(-fraction) %>% 
  rename(Netherlands = fraction) %>% 
  right_join(fraction_data, by = "country"); rm(temp)

fraction_data
```

```{r}

heatmap_df <- fraction_data %>% 
  mutate(country = recode(country, "United States" = "USA", "United Kingdom" = "UK")) %>% 
  arrange(match(country, c("Australia", "Canada", "China", "France", "Germany", "India", "Japan", "Netherlands", "UK", "USA"))) %>% 
  #select(country, Australia, Canada, China, France, Germany, India, Japan, Netherlands, UK, USA) %>% 
  select(country, USA, UK, Netherlands, Japan, India, Germany, France, China, Canada, Australia) %>% 
  rename(Country = country)
heatmap_df

heatmap_matrix <- as.matrix(fraction_data %>% 
  mutate(country = recode(country, "United States" = "USA", "United Kingdom" = "UK")) %>% 
  arrange(match(country, c("Australia", "Canada", "China", "France", "Germany", "India", "Japan", "Netherlands", "UK", "USA"))) %>% 
  select(country, Australia, Canada, China, France, Germany, India, Japan, Netherlands, UK, USA) %>% 
  column_to_rownames(var = "country"))
t(heatmap_matrix)
```

```{r}

melted_for_heatmap <- melt(heatmap_df)

melted_for_heatmap %>% 
  mutate(Country = fct_relevel(Country, "Australia", "Canada", "China", "France", "Germany", "India", "Japan", "Netherlands", "UK", "USA")) %>% 
  #mutate(variable = fct_relevel(variable, "Australia", "Canada", "China", "France", "Germany", "India", "Japan", "Netherlands", "UK", "USA")) %>%
  ggplot(aes(Country, variable, fill= value)) + 
  geom_tile() +
  scale_fill_distiller(palette = "Spectral") +
  theme(axis.title.y=element_blank(), #legend.position="bottom",
        #axis.text.x = element_text(angle=20, hjust = 1),
        #axis.text.y = element_text(angle=20, hjust = 1),
        plot.title = element_text(size=15, hjust =0.46)) +
  scale_x_discrete(name = "Fraction of Total Collaborations") + 
  ggtitle("Fraction of Total Collaborations with \n Other Top-10 Countries (GitHub, 2019)") +
  guides(fill=guide_legend(title="Fraction"))

```
